# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "project setup API",
        "contact": {
            "name": "2F Capital Support Email",
            "url": "http://www.2fcapital.com",
            "email": "info@1f-capital.com"
        },
        "version": "0.1"
    },
    "host": "localhost:8000",
    "basePath": "/v1",
    "paths": {
        "/users": {
            "get": {
                "description": "This endpoint is used to retrieve a list of users.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get users.",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, check the error response for details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ErrorResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint is used to create a new user by providing the necessary details in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a new user.",
                "parameters": [
                    {
                        "description": "User details in JSON format",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created user",
                        "schema": {
                            "$ref": "#/definitions/dto.User"
                        }
                    },
                    "400": {
                        "description": "Bad request, check the error response for details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ErrorResponse"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "This endpoint is used to retrieve information about a user identified by the provided User ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user",
                        "schema": {
                            "$ref": "#/definitions/dto.User"
                        }
                    },
                    "400": {
                        "description": "Bad request, check the error response for details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ErrorResponse"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "This function deletes a user if the user is available.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ErrorResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ErrorResponse"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "This endpoint is used to update an existing user identified by the provided User ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update an existing user.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated user details in JSON format",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/dto.User"
                        }
                    },
                    "400": {
                        "description": "Bad request, check the error response for details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ErrorResponse"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.RegisterUser": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email is the email of the user.",
                    "type": "string"
                },
                "first_name": {
                    "description": "FirstName is the first name of the user.",
                    "type": "string"
                },
                "last_name": {
                    "description": "LastName is the last name of the user.",
                    "type": "string"
                },
                "middle_name": {
                    "description": "MiddleName is the middle name of the user.",
                    "type": "string"
                }
            }
        },
        "dto.UpdateUser": {
            "type": "object",
            "properties": {
                "first_name": {
                    "description": "FirstName is the first name of the user.",
                    "type": "string"
                },
                "last_name": {
                    "description": "LastName is the last name of the user.",
                    "type": "string"
                },
                "middle_name": {
                    "description": "MiddleName is the middle name of the user.",
                    "type": "string"
                }
            }
        },
        "dto.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt is the time when the user is created.\nIt is automatically set when the user is created.",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "DeletedAt is the time the user was deleted.",
                    "type": "string"
                },
                "email": {
                    "description": "Email is the email of the user.",
                    "type": "string"
                },
                "first_name": {
                    "description": "FirstName is the first name of the user.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the unique identifier of the user.\nIt is automatically generated when the user is created.",
                    "type": "string"
                },
                "last_name": {
                    "description": "LastName is the last name of the user.",
                    "type": "string"
                },
                "middle_name": {
                    "description": "MiddleName is the middle name of the user.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the status of the user.\nIt is set to active by default after successful registration.",
                    "type": "string"
                },
                "updated_at": {
                    "description": "UpdatedAt is the time the user was last updated.",
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the error code. It is not status code",
                    "type": "integer"
                },
                "description": {
                    "description": "Description is the error description.",
                    "type": "string"
                },
                "field_error": {
                    "description": "FieldError is the error detail for each field, if available that is.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FieldError"
                    }
                },
                "message": {
                    "description": "Message is the error message.",
                    "type": "string"
                },
                "stack_trace": {
                    "description": "StackTrace is the stack trace of the error.\nIt is only returned for debugging",
                    "type": "string"
                }
            }
        },
        "model.FieldError": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description is the error description for this field.",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the name of the field that caused the error.",
                    "type": "string"
                }
            }
        }
    }
}